moved serverFast to default server.mjs
